requirements.txt
streamlit
requests
pycountry
altair
python-dotenv
Babel